refactor(Profile) : profile/ui Modal 폴더 변수명 수정 컨벤션에 따라 수식어 명사 형태로 수정
