+++ src/1_Page/Profile/ui/ModifyModeModal/index.js
@@ -3,7 +3,7 @@ import Modal from "../../../../2_Widget/Modal";
 import STYLE from "./style";
 import ModalConfirm from "../../../../2_Widget/ModalConfirm";
 
-const ModalModifyMode = (props) => {
+const ModifyModeModal = (props) => {
   const { handleModifyClickFalse, handleSetMode, sumDataLength } = props;
 
   const handleShare = () => {
@@ -50,4 +50,4 @@ const ModalModifyMode = (props) => {
   );
 };
 
-export default ModalModifyMode;
+export default ModifyModeModal;
